# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AfraLink transportation and logistics marketplace API for Southern Nigeria
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      operationId: getMyProfile
      summary: Get current user's full profile
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: User profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [Users]
      operationId: updateMyProfile
      summary: Update current user's profile
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdate'
      responses:
        '200':
          description: Updated profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  # ─── DRIVERS ─────────────────────────────────────────────────────────────────
  /drivers:
    get:
      tags: [Drivers]
      operationId: listDrivers
      summary: List and search available drivers
      parameters:
        - name: state
          in: query
          required: false
          schema: { type: string }
        - name: city
          in: query
          required: false
          schema: { type: string }
        - name: vehicleType
          in: query
          required: false
          schema: { type: string }
        - name: page
          in: query
          required: false
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: List of drivers.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverListResponse'
    post:
      tags: [Drivers]
      operationId: createDriver
      summary: Register a driver profile
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverInput'
      responses:
        '201':
          description: Driver profile created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /drivers/{id}:
    get:
      tags: [Drivers]
      operationId: getDriver
      summary: Get a driver profile by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Driver profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [Drivers]
      operationId: updateDriver
      summary: Update a driver profile
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverUpdate'
      responses:
        '200':
          description: Updated driver.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /drivers/me:
    get:
      tags: [Drivers]
      operationId: getMyDriver
      summary: Get the current user's driver profile
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Current user's driver profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
        '404':
          description: No driver profile found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  # ─── CAR RENTALS ─────────────────────────────────────────────────────────────
  /rentals:
    get:
      tags: [Rentals]
      operationId: listRentals
      summary: List and search rental vehicles
      parameters:
        - name: state
          in: query
          required: false
          schema: { type: string }
        - name: city
          in: query
          required: false
          schema: { type: string }
        - name: vehicleType
          in: query
          required: false
          schema: { type: string }
        - name: page
          in: query
          required: false
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: List of rental vehicles.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RentalListResponse'
    post:
      tags: [Rentals]
      operationId: createRental
      summary: List a vehicle for rental
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RentalInput'
      responses:
        '201':
          description: Rental listing created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rental'
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /rentals/{id}:
    get:
      tags: [Rentals]
      operationId: getRental
      summary: Get a rental listing by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Rental listing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rental'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [Rentals]
      operationId: updateRental
      summary: Update a rental listing
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RentalUpdate'
      responses:
        '200':
          description: Updated rental.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rental'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Rentals]
      operationId: deleteRental
      summary: Remove a rental listing
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '204':
          description: Deleted.
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  # ─── BOOKINGS ─────────────────────────────────────────────────────────────────
...
  /reviews:
    get:
      tags: [Reviews]
      operationId: listReviews
      summary: List reviews for a driver or rental
      parameters:
        - name: driverId
          in: query
          required: false
          schema: { type: integer }
        - name: rentalId
          in: query
          required: false
          schema: { type: integer }
      responses:
        '200':
          description: List of reviews.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewListResponse'
    post:
      tags: [Reviews]
      operationId: createReview
      summary: Submit a rating and review
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewInput'
      responses:
        '201':
          description: Review created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  # ─── REPORTS ─────────────────────────────────────────────────────────────────
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema: { type: string }
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }

    # ── STORAGE SCHEMAS ─────────────────────────────────────────────────────────
...
    UserProfile:
      type: object
      required: [id, role, createdAt]
      properties:
        id: { type: string }
        role: { type: string, enum: [customer, driver, rental_owner, admin] }
        fullName: { type: [string, 'null'] }
        phone: { type: [string, 'null'] }
        email: { type: [string, 'null'] }
        profilePhotoUrl: { type: [string, 'null'] }
        state: { type: [string, 'null'] }
        city: { type: [string, 'null'] }
        isSuspended: { type: boolean }
        createdAt: { type: string, format: date-time }
    UserProfileUpdate:
      type: object
      properties:
        fullName: { type: string }
        phone: { type: string }
        profilePhotoUrl: { type: string }
        state: { type: string }
        city: { type: string }
        role: { type: string, enum: [customer, driver, rental_owner] }

    # ── DRIVER SCHEMAS ──────────────────────────────────────────────────────────
    Driver:
      type: object
      required: [id, userId, fullName, phone, vehicleType, state, city, verificationStatus, isAvailable, createdAt]
      properties:
        id: { type: integer }
        userId: { type: string }
        fullName: { type: string }
        phone: { type: string }
        email: { type: [string, 'null'] }
        gender: { type: [string, 'null'] }
        dateOfBirth: { type: [string, 'null'] }
        address: { type: [string, 'null'] }
        state: { type: string }
        city: { type: string }
        vehicleType: { type: string }
        vehicleBrand: { type: [string, 'null'] }
        vehicleModel: { type: [string, 'null'] }
        vehicleColor: { type: [string, 'null'] }
        plateNumber: { type: [string, 'null'] }
        ninNumber: { type: [string, 'null'] }
        ninSlipUrl: { type: [string, 'null'] }
        selfieUrl: { type: [string, 'null'] }
        vehiclePhotoUrl: { type: [string, 'null'] }
        profilePhotoUrl: { type: [string, 'null'] }
        verificationStatus: { type: string, enum: [pending, approved, rejected] }
        rejectionNote: { type: [string, 'null'] }
        isAvailable: { type: boolean }
        averageRating: { type: [number, 'null'] }
        reviewCount: { type: integer }
        isFeatured: { type: boolean }
        createdAt: { type: string, format: date-time }
    DriverInput:
      type: object
      required: [fullName, phone, state, city, vehicleType]
      properties:
        fullName: { type: string }
        phone: { type: string }
        email: { type: string }
        gender: { type: string }
        dateOfBirth: { type: string }
        address: { type: string }
        state: { type: string }
        city: { type: string }
        vehicleType: { type: string }
        vehicleBrand: { type: string }
        vehicleModel: { type: string }
        vehicleColor: { type: string }
        plateNumber: { type: string }
        ninNumber: { type: string }
        ninSlipUrl: { type: string }
        selfieUrl: { type: string }
        vehiclePhotoUrl: { type: string }
        profilePhotoUrl: { type: string }
    DriverUpdate:
      type: object
      properties:
        fullName: { type: string }
        phone: { type: string }
        vehicleBrand: { type: string }
        vehicleModel: { type: string }
        vehicleColor: { type: string }
        plateNumber: { type: string }
        ninSlipUrl: { type: string }
        selfieUrl: { type: string }
        vehiclePhotoUrl: { type: string }
        profilePhotoUrl: { type: string }
        isAvailable: { type: boolean }
        state: { type: string }
        city: { type: string }
    DriverListResponse:
      type: object
      required: [drivers, total]
      properties:
        drivers:
          type: array
          items:
            $ref: '#/components/schemas/Driver'
        total: { type: integer }

    # ── RENTAL SCHEMAS ──────────────────────────────────────────────────────────
    Rental:
      type: object
      required: [id, userId, vehicleName, brand, model, state, city, dailyPrice, verificationStatus, isAvailable, createdAt]
      properties:
        id: { type: integer }
        userId: { type: string }
        ownerName: { type: [string, 'null'] }
        ownerPhone: { type: [string, 'null'] }
        vehicleName: { type: string }
        brand: { type: string }
        model: { type: string }
        year: { type: [integer, 'null'] }
        color: { type: [string, 'null'] }
        vehicleType: { type: [string, 'null'] }
        transmission: { type: [string, 'null'] }
        fuelType: { type: [string, 'null'] }
        seatingCapacity: { type: [integer, 'null'] }
        plateNumber: { type: [string, 'null'] }
        state: { type: string }
        city: { type: string }
        dailyPrice: { type: number }
        weeklyPrice: { type: [number, 'null'] }
        monthlyPrice: { type: [number, 'null'] }
        photoUrls: { type: array, items: { type: string } }
        verificationStatus: { type: string, enum: [pending, approved, rejected] }
        isAvailable: { type: boolean }
        isFeatured: { type: boolean }
        averageRating: { type: [number, 'null'] }
        reviewCount: { type: integer }
        createdAt: { type: string, format: date-time }
    RentalInput:
      type: object
      required: [vehicleName, brand, model, state, city, dailyPrice]
      properties:
        vehicleName: { type: string }
        brand: { type: string }
        model: { type: string }
        year: { type: integer }
        color: { type: string }
        vehicleType: { type: string }
        transmission: { type: string }
        fuelType: { type: string }
        seatingCapacity: { type: integer }
        plateNumber: { type: string }
        state: { type: string }
        city: { type: string }
        dailyPrice: { type: number }
        weeklyPrice: { type: number }
        monthlyPrice: { type: number }
        photoUrls: { type: array, items: { type: string } }
        ownerPhone: { type: string }
    RentalUpdate:
      type: object
      properties:
        vehicleName: { type: string }
        dailyPrice: { type: number }
        weeklyPrice: { type: number }
        monthlyPrice: { type: number }
        isAvailable: { type: boolean }
        photoUrls: { type: array, items: { type: string } }
        ownerPhone: { type: string }
    RentalListResponse:
      type: object
      required: [rentals, total]
      properties:
        rentals:
          type: array
          items:
            $ref: '#/components/schemas/Rental'
        total: { type: integer }

    # ── BOOKING SCHEMAS ─────────────────────────────────────────────────────────
...
    Review:
      type: object
      required: [id, reviewerId, rating, createdAt]
      properties:
        id: { type: integer }
        reviewerId: { type: string }
        reviewerName: { type: [string, 'null'] }
        driverId: { type: [integer, 'null'] }
        rentalId: { type: [integer, 'null'] }
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
    ReviewInput:
      type: object
      required: [rating]
      properties:
        driverId: { type: integer }
        rentalId: { type: integer }
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string }
    ReviewListResponse:
      type: object
      required: [reviews]
      properties:
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'

    # ── REPORT SCHEMAS ──────────────────────────────────────────────────────────
...