# lib/api-zod/src/generated/types/rentalInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AfraLink transportation and logistics marketplace API for Southern Nigeria
 * OpenAPI spec version: 0.1.0
 */

export interface RentalInput {
  vehicleName: string;
  brand: string;
  model: string;
  year?: number;
  color?: string;
  vehicleType?: string;
  transmission?: string;
  fuelType?: string;
  seatingCapacity?: number;
  plateNumber?: string;
  state: string;
  city: string;
  dailyPrice: number;
  weeklyPrice?: number;
  monthlyPrice?: number;
  photoUrls?: string[];
  ownerPhone?: string;
}