# lib/db/src/schema/reviews.ts
import { pgTable, text, integer, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const reviewsTable = pgTable("reviews", {
  id: integer("id").primaryKey().generatedAlwaysAsIdentity(),
  reviewerId: text("reviewer_id").notNull(),
  reviewerName: text("reviewer_name"),
  driverId: integer("driver_id"),
  rentalId: integer("rental_id"),
  rating: integer("rating").notNull(),
  comment: text("comment"),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
});

export const insertReviewSchema = createInsertSchema(reviewsTable).omit({ createdAt: true });
export type InsertReview = z.infer<typeof insertReviewSchema>;
export type Review = typeof reviewsTable.$inferSelect;